Added Manage Profile

+++ EcommMVC/Controllers/OrderController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Data.Entity.Migrations;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Web;
@@ -30,13 +31,13 @@ namespace EcommMVC.Controllers
         public ActionResult Index()
         {
 
-           var myOrders =  _context.OrderDetails.ToList().Where(o => o.UserId == User.Identity.GetUserId());
+            var myOrders = _context.OrderDetails.ToList().Where(o => o.UserId == User.Identity.GetUserId());
 
             return View(myOrders);
         }
 
 
-        // GET: /Orders/PlaceOrder
+        // GET: /Order/PlaceOrder
         public ActionResult PlaceOrder()
         {
 
@@ -44,7 +45,9 @@ namespace EcommMVC.Controllers
 
             var check = true;
             string OrderId = "";
-
+            var userId = User.Identity.GetUserId();
+            var CurrentUser = _context.Users.FirstOrDefault(u => u.Id == userId);
+            var cartItems = _context.Carts.ToList().Where(c => c.UserId == userId);
             while (check)
             {
                 var guid = Guid.NewGuid().ToString();
@@ -56,16 +59,28 @@ namespace EcommMVC.Controllers
                 {
                     check = false;
                     OrderId = guid;
+                    Session["OrderId"] = guid;
                 }
 
             }
 
-            Session["OrderId"] = OrderId;
 
-            return View();
+            var order = new Order()
+            {
+                OrderId = OrderId,
+                UserId = User.Identity.GetUserId(),
+                Email = CurrentUser.Email,
+                SaveInfo = true,
+
+            };
+
+            Session["cartList"] = cartItems;
+            Session["WalletBalance"] = CurrentUser.Wallet;
+
+            return View(order);
         }
 
-        // POST: /Orders/PlaceOrder
+        // POST: /Order/PlaceOrder
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> PlaceOrder(Order order)
@@ -77,41 +92,43 @@ namespace EcommMVC.Controllers
                 return View(order);
             }
 
-
+            var userId = User.Identity.GetUserId();
             var TotalPayable = (double)Session["totalPayable"];
-            var cart = _context.Carts.ToList().Where(c => c.UserId == User.Identity.GetUserId());
-            var orderDetails = (IEnumerable<OrderDetails>)Session["orderList"];
-            //var CurrentUser = _context.Users.SingleOrDefault(u => u.Id == User.Identity.GetUserId());
-
-
-            //var Balance = CurrentUser.Wallet - TotalPayable;
-
-
-            foreach (var item in orderDetails)
-            {
-
-                item.OrderId = order.OrderId;
-                _context.OrderDetails.Add(item);
-
-
-            }
-
-            foreach (var item in cart)
-            {
-                _context.Carts.Remove(item);
-            }
-
+            var CurrentUser = _context.Users.SingleOrDefault(u => u.Id == userId);
+            var Balance = CurrentUser.Wallet - TotalPayable;
 
             _context.Orders.Add(order);
 
+            if (Balance < 0.0)
+            {
+                return View(order);
+            }
 
+            var cart = _context.Carts.ToList().Where(c => c.UserId == userId);
+
+            foreach (var item in cart)
+            {
+                _context.OrderDetails.Add(new OrderDetails()
+                {
+                    OrderId = order.OrderId,
+                    ProductId = item.ItemId,
+                    UserId = item.UserId,
+                    ItemQuantity = item.ItemQuantity,
+                    ItemName = item.ItemName,
+                    TotalPrice = item.ItemPrice * item.ItemQuantity,
+                    OrderStatus = "Order Placed",
+                    DeliverStatus = false,
+                });
+
+                _context.Carts.Remove(item);
+            }
+
+            CurrentUser.Wallet = Balance;
+            _context.Users.AddOrUpdate(CurrentUser);
 
             UpdateDatabase();
 
-
             return RedirectToAction("Index", "ProductDetails");
-
-
 
         }
